Implement DataParser class
DataParser class implemented with cooresponding unit tests. This class retrieves the data from the Google Spreadsheet and then parses it into individual objects for each of the responses in the spreadsheet. Also implemented the equals method for the Person class, moved the acceptance tests to their own folder, and modified some method signatures and JavaDoc.