Change AT files
I modified how I plan to generate the files. Changed the test files accordingly. I'm using fewer tables to reduce the amount of whitespace between data entries and as well to limit the amount of css needed. I also added a `style.css` file to the tests, which is to ensure the files display correctly.