Implemented FileGenerator
FileGenerator class first creates a directory for the files, adds the `style.css`, then creates a file for each student in the spreadsheet data given. Added some functionality to the Person class and added a unit test for when PreferredName is null (a bug which was caught while generating the files in unit tests of FileGenerator).