1.0 Release
Modified clean-up in ATs to be more bullet proof and ensure we don't get false postives. Also added px to the css attributes, since they weren't working properly otherwise. This release produces a CLI client that can be used to generate the one-page view for each student. There are currently no customizations on the view of each file, and the UI is non-existent at the moment (hopefully additions in future releases.