Handle URL with "edit#gid" at the end
Previously an error was thrown when this part of the url was included. It's now correctly stripped off the url when the spreadsheetId is extracted. This was changed since most end users will likely just copy paste the url from their browser which typically contains this string at the end.